improve code and adding tailwindcss classes on typing
