+++ types/index.ts
@@ -42,3 +42,5 @@ export type ColorSwatchProps = {
   name: string;
   size?: "sm" | "md" | "lg";
 };
+
+export type ValidShade = '50' | '100' | '200' | '300' | '400' | '500' | '600' | '700' | '800' | '900' | '950';
